+++ quantify_scheduler/backends/corrections.py
@@ -173,8 +173,8 @@ def apply_distortion_corrections(
     pulse. This means that we can have a combination of corrected (i.e., pre-sampled)
     and uncorrected pulses in the same operation.
 
-    Note that we are **not** updating the ``"operation_repr"`` key, used to reference
-    the operation from the schedulable.
+    Note that we are **not** updating the ``"operation_id"`` key, used to reference
+    the operation from schedulables.
 
     Parameters
     ----------
@@ -207,22 +207,22 @@ def apply_distortion_corrections(
         logging.debug(f'No key "{distortion_corrections_key}" supplied in hardware_cfg')
         return schedule
 
-    for operation_repr in schedule.operations.keys():
+    for operation_id in schedule.operations.keys():
         substitute_operation = None
 
         for pulse_info_idx, pulse_data in enumerate(
-            schedule.operations[operation_repr].data["pulse_info"]
+            schedule.operations[operation_id].data["pulse_info"]
         ):
             portclock_key = f"{pulse_data['port']}-{pulse_data['clock']}"
 
             if portclock_key in hardware_cfg[distortion_corrections_key]:
-                if not _is_distortion_correctable(schedule.operations[operation_repr]):
+                if not _is_distortion_correctable(schedule.operations[operation_id]):
                     warnings.warn(
                         f"Schedule contains an operation, for which distortion "
                         f"correction is not implemented. Please either replace the "
                         f"operation, or omit the distortion correction setting for "
                         f"this port in order to suppress this warning. Offending "
-                        f"operation: {schedule.operations[operation_repr]}",
+                        f"operation: {schedule.operations[operation_id]}",
                         RuntimeWarning,
                     )
                     continue
@@ -259,7 +259,7 @@ def apply_distortion_corrections(
                     clipping_values=clipping_values,
                 )
 
-                schedule.operations[operation_repr].data["pulse_info"][
+                schedule.operations[operation_id].data["pulse_info"][
                     pulse_info_idx
                 ] = corrected_pulse.data["pulse_info"][0]
 
@@ -270,8 +270,8 @@ def apply_distortion_corrections(
         # required as first entry in pulse_info is used to generate signature in __str__
         if substitute_operation is not None:
             substitute_operation.data["pulse_info"] = schedule.operations[
-                operation_repr
+                operation_id
             ].data["pulse_info"]
-            schedule.operations[operation_repr] = substitute_operation
+            schedule.operations[operation_id] = substitute_operation
 
     return schedule