+++ tests/scheduler/helpers/test_schedule.py
@@ -68,10 +68,10 @@ def test_get_info_by_uuid(
     schedule_with_pulse_info: Schedule,
 ):
     # Arrange
-    operation_repr = list(schedule_with_pulse_info.schedulables.values())[0][
-        "operation_repr"
+    operation_id = list(schedule_with_pulse_info.schedulables.values())[0][
+        "operation_id"
     ]
-    pulse_info_0 = schedule_with_pulse_info.operations[operation_repr]["pulse_info"][0]
+    pulse_info_0 = schedule_with_pulse_info.operations[operation_id]["pulse_info"][0]
     pulse_id = get_pulse_uuid(pulse_info_0)
 
     # Act
@@ -94,8 +94,8 @@ def test_get_info_by_uuid_are_unique(device_compile_config_basic_transmon):
         schedule=schedule, config=device_compile_config_basic_transmon
     )
 
-    operation_repr = list(schedule.schedulables.values())[0]["operation_repr"]
-    pulse_info_0 = schedule_with_pulse_info.operations[operation_repr]["pulse_info"][0]
+    operation_id = list(schedule.schedulables.values())[0]["operation_id"]
+    pulse_info_0 = schedule_with_pulse_info.operations[operation_id]["pulse_info"][0]
     pulse_id = get_pulse_uuid(pulse_info_0)
 
     # Act
@@ -117,8 +117,8 @@ def test_get_acq_info_by_uuid(
         schedule=schedule_with_measurement, config=device_compile_config_basic_transmon
     )
 
-    operation_repr = list(schedule.schedulables.values())[-1]["operation_repr"]
-    operation = schedule.operations[operation_repr]
+    operation_id = list(schedule.schedulables.values())[-1]["operation_id"]
+    operation = schedule.operations[operation_id]
     acq_info_0 = operation["acquisition_info"][0]
     acq_pulse_infos = acq_info_0["waveforms"]
 
@@ -141,10 +141,10 @@ def test_get_port_timeline(
     schedule_with_pulse_info: Schedule,
 ):
     # Arrange
-    operation_repr = list(schedule_with_pulse_info.schedulables.values())[0][
-        "operation_repr"
+    operation_id = list(schedule_with_pulse_info.schedulables.values())[0][
+        "operation_id"
     ]
-    pulse_info_0 = schedule_with_pulse_info.operations[operation_repr]["pulse_info"][0]
+    pulse_info_0 = schedule_with_pulse_info.operations[operation_id]["pulse_info"][0]
     pulse_id = get_pulse_uuid(pulse_info_0)
     port = pulse_info_0["port"]
     timeslot_index = 0
@@ -188,19 +188,19 @@ def test_get_port_timeline_sorted(
         schedule=schedule, config=device_compile_config_basic_transmon
     )
 
-    reset_operation_id = list(schedule.schedulables.values())[0]["operation_repr"]
+    reset_operation_id = list(schedule.schedulables.values())[0]["operation_id"]
     reset_pulse_info = schedule.operations[reset_operation_id]["pulse_info"][0]
     reset_pulse_id = get_pulse_uuid(reset_pulse_info)
 
-    qubit_operation_id = list(schedule.schedulables.values())[1]["operation_repr"]
+    qubit_operation_id = list(schedule.schedulables.values())[1]["operation_id"]
     qubit_pulse_info = schedule.operations[qubit_operation_id]["pulse_info"][0]
     qubit_pulse_id = get_pulse_uuid(qubit_pulse_info)
 
-    ro_operation_id = list(schedule.schedulables.values())[2]["operation_repr"]
+    ro_operation_id = list(schedule.schedulables.values())[2]["operation_id"]
     ro_pulse_info = schedule.operations[ro_operation_id]["pulse_info"][0]
     ro_pulse_id = get_pulse_uuid(ro_pulse_info)
 
-    acq_operation_id = list(schedule.schedulables.values())[3]["operation_repr"]
+    acq_operation_id = list(schedule.schedulables.values())[3]["operation_id"]
     acq_pulse_info = schedule.operations[acq_operation_id]["acquisition_info"][0]
     acq_id = get_acq_uuid(acq_pulse_info)
 
@@ -243,17 +243,17 @@ def test_get_port_timeline_are_unique(device_compile_config_basic_transmon):
         schedule=schedule, config=device_compile_config_basic_transmon
     )
 
-    reset_operation_id = list(schedule.schedulables.values())[0]["operation_repr"]
+    reset_operation_id = list(schedule.schedulables.values())[0]["operation_id"]
     reset_pulse_info_q0 = schedule.operations[reset_operation_id]["pulse_info"][0]
     reset_pulse_id_q0 = get_pulse_uuid(reset_pulse_info_q0)
     reset_pulse_info_q1 = schedule.operations[reset_operation_id]["pulse_info"][1]
     reset_pulse_id_q1 = get_pulse_uuid(reset_pulse_info_q1)
 
-    q0_operation_id = list(schedule.schedulables.values())[1]["operation_repr"]
+    q0_operation_id = list(schedule.schedulables.values())[1]["operation_id"]
     q0_pulse_info = schedule.operations[q0_operation_id]["pulse_info"][0]
     q0_pulse_id = get_pulse_uuid(q0_pulse_info)
 
-    q1_operation_id = list(schedule.schedulables.values())[2]["operation_repr"]
+    q1_operation_id = list(schedule.schedulables.values())[2]["operation_id"]
     q1_pulse_info = schedule.operations[q1_operation_id]["pulse_info"][0]
     q1_pulse_id = get_pulse_uuid(q1_pulse_info)
 
@@ -284,7 +284,7 @@ def test_get_port_timeline_with_duplicate_op(device_compile_config_basic_transmo
         schedule=schedule, config=device_compile_config_basic_transmon
     )
 
-    X90_q0_operation_id = list(schedule.schedulables.values())[0]["operation_repr"]
+    X90_q0_operation_id = list(schedule.schedulables.values())[0]["operation_id"]
     X90_q0_pulse_info = schedule.operations[X90_q0_operation_id]["pulse_info"][0]
     X90_q0_pulse_id = get_pulse_uuid(X90_q0_pulse_info)
 
@@ -310,17 +310,17 @@ def test_get_port_timeline_with_acquisition(
 
     schedule = create_schedule_with_pulse_info(schedule_with_measurement, device_config)
 
-    reset_operation_id = list(schedule.schedulables.values())[0]["operation_repr"]
+    reset_operation_id = list(schedule.schedulables.values())[0]["operation_id"]
     reset_operation = schedule.operations[reset_operation_id]
     reset_pulse_info = reset_operation["pulse_info"][0]
     reset_pulse_id = get_pulse_uuid(reset_pulse_info)
 
-    q0_operation_id = list(schedule.schedulables.values())[1]["operation_repr"]
+    q0_operation_id = list(schedule.schedulables.values())[1]["operation_id"]
     q0_operation = schedule.operations[q0_operation_id]
     q0_pulse_info = q0_operation["pulse_info"][0]
     q0_pulse_id = get_pulse_uuid(q0_pulse_info)
 
-    acq_operation_id = list(schedule.schedulables.values())[2]["operation_repr"]
+    acq_operation_id = list(schedule.schedulables.values())[2]["operation_id"]
     acq_operation = schedule.operations[acq_operation_id]
 
     # Acquisition consists of a reset_clock_phase instruction