+++ tests/scheduler/helpers/test_waveforms.py
@@ -88,10 +88,10 @@ def test_get_waveform_by_pulseid(
     schedule_with_pulse_info: Schedule,
 ) -> None:
     # Arrange
-    operation_repr = list(schedule_with_pulse_info.schedulables.values())[0][
-        "operation_repr"
+    operation_id = list(schedule_with_pulse_info.schedulables.values())[0][
+        "operation_id"
     ]
-    pulse_info_0 = schedule_with_pulse_info.operations[operation_repr]["pulse_info"][0]
+    pulse_info_0 = schedule_with_pulse_info.operations[operation_id]["pulse_info"][0]
     pulse_id = get_pulse_uuid(pulse_info_0)
     expected_keys: List[int] = [pulse_id]
 
@@ -117,8 +117,8 @@ def test_get_waveform_by_pulseid_are_unique(
         schedule=schedule, config=device_compile_config_basic_transmon
     )
 
-    operation_repr = list(schedule.schedulables.values())[0]["operation_repr"]
-    pulse_info_0 = schedule.operations[operation_repr]["pulse_info"][0]
+    operation_id = list(schedule.schedulables.values())[0]["operation_id"]
+    pulse_info_0 = schedule.operations[operation_id]["pulse_info"][0]
     pulse_id = get_pulse_uuid(pulse_info_0)
     expected_keys: List[int] = [pulse_id]