+++ tests/scheduler/schedules/test_timedomain_schedules.py
@@ -109,9 +109,7 @@ class TestRabiSched(_CompilesAllBackends):
             assert schedulable["abs_time"] == abs_times[i]
 
     def test_rabi_pulse_ops(self):
-        rabi_op_hash = list(self.uncomp_sched.schedulables.values())[1][
-            "operation_repr"
-        ]
+        rabi_op_hash = list(self.uncomp_sched.schedulables.values())[1]["operation_id"]
         rabi_pulse = self.uncomp_sched.operations[rabi_op_hash]["pulse_info"][0]
         assert rabi_pulse["G_amp"] == 0.2
         assert rabi_pulse["D_amp"] == 0
@@ -138,7 +136,7 @@ class TestRabiSched(_CompilesAllBackends):
         for i, schedulable in enumerate(sched.schedulables.values()):
             assert schedulable["label"] == labels[i]
 
-        rabi_op_hash = list(sched.schedulables.values())[1]["operation_repr"]
+        rabi_op_hash = list(sched.schedulables.values())[1]["operation_id"]
         rabi_pulse = sched.operations[rabi_op_hash]["pulse_info"][0]
         assert rabi_pulse["G_amp"] == 0.5
         assert rabi_pulse["D_amp"] == 0
@@ -170,9 +168,7 @@ class TestRabiSched(_CompilesAllBackends):
             assert schedulable["label"] == labels[i]
 
         for i, exp_amp in enumerate(amps):
-            rabi_op_hash = list(sched.schedulables.values())[3 * i + 1][
-                "operation_repr"
-            ]
+            rabi_op_hash = list(sched.schedulables.values())[3 * i + 1]["operation_id"]
             rabi_pulse = sched.operations[rabi_op_hash]["pulse_info"][0]
             assert rabi_pulse["G_amp"] == exp_amp
             assert rabi_pulse["D_amp"] == 0
@@ -217,7 +213,7 @@ class TestRabiSched(_CompilesAllBackends):
 
     def test_correct_inference_of_port_clock(self):
         # operation 1 is tested in test_timing to be the Rabi pulse
-        op_name = list(self.uncomp_sched.schedulables.values())[1]["operation_repr"]
+        op_name = list(self.uncomp_sched.schedulables.values())[1]["operation_id"]
         rabi_op = self.uncomp_sched.operations[op_name]
         assert rabi_op["pulse_info"][0]["port"] == "q0:mw"
         assert rabi_op["pulse_info"][0]["clock"] == "q0.01"
@@ -316,7 +312,7 @@ class TestCPMGSched(_CompilesAllBackends):
                 sub_sched = self.uncomp_sched.operations[key]
                 sub_sched_duration = 0
                 for sub_schedulable in sub_sched.schedulables.values():
-                    operation = sub_sched.operations[sub_schedulable["operation_repr"]]
+                    operation = sub_sched.operations[sub_schedulable["operation_id"]]
                     if operation["name"] == "IdlePulse":
                         sub_sched_duration += operation["pulse_info"][0]["duration"]
                 n_reps = self.sched_kwargs["n_gates"]